Refs #36878 - move action to new controller
