+++ lib/katello/engine.rb
@@ -197,6 +197,7 @@ module Katello
       #Api controller extensions
       ::Api::V2::HostsController.include Katello::Concerns::Api::V2::HostsControllerExtensions
       ::Api::V2::HostsController.include Katello::Concerns::ContentFacetHostsControllerExtensions
+      ::Api::V2::HostsBulkActionsController.include Katello::Concerns::Api::V2::HostsBulkActionsControllerExtensions
       ::Api::V2::HostgroupsController.include Katello::Concerns::Api::V2::HostgroupsControllerExtensions
       ::Api::V2::SmartProxiesController.include Katello::Concerns::Api::V2::SmartProxiesControllerExtensions
       ::Api::V2::RegistrationController.include ::Foreman::Controller::SmartProxyAuth
